requirements.txt
pyTelegramBotAPI==4.12.0
supabase>=2.28.0
python-dotenv==1.0.0
requests==2.28.2
